ENH: Adds Threshold button, pushes threshold label to Slicer
This adds a threshold button that uses the SimpleITK BinaryThreshold
filter with a lower threshold value of 130.  The result is then
type casted into a short sitk.sitkInt16 image as required by the
Editor module.  This label is pushed to Slicer and will be used to
create 4 label maps of different calcium plaques using the
ChangeIslandEffect in the Editor Module.