ENH: Moves the creation of EditorWidget to AFTER colorNode set
I noticed that if I reloaded the module, then the Editor Widget
would work as expected - it would have loaded the color lookup
table for use in editing. So I moved the creation and setup of
the Editor table to after the threshold image was set to the
label and the custom lookup table was asigned to the colorNode.
This works!