ENH: Removed grayscaleNode and labelNode since they are set to active volumes
I removed the code for selecting the grayscaleNode and labelNode
since I set the grayscaleNode and labelNode to the current active
volume and label.  This reduces the number of steps the user must
take to get the label statistics, but the user must have the
correct background and label image set. These are set by default.