ENH: Removes ChangeIslandEffect tool, moves label buttons
The ChangeIslandEffect tool was removed because the four
specific label buttons now select the ChangeIslandEffect
and change the label color.  Therefore, it is not needed
anymore.  The four label buttons were moved from the parent
layout to the mainFrame layout.  Then the DefaultTool was
added to the button row for Undo/Redo.