ENH: Calculates score for each island now, not just slice and label
The Agatston Score is calculated by summing the scores for each
island. We were calculating the score for each slice and then
summing those values.  This change is important because each
island within a slice could have significantly diffent maximum
intensity values and therefore different x-factors (1,2,3, or 4).
Before we were selecting the maximum intesity value for the entire
slice/label and multiplying that x-factor by a summed area.