ENH: Adds test and self-test code created using ModuleWizard
Added classes CardiacAgatstonMeasuresLogic and
CardiacAgatstonMeasuresTest created using the ModuleWizard
found at http://www.slicer.org/slicerWiki/index.php/
Documentation/Nightly/Developers/ModuleWizard. This code
adds a default self-test to the Slicer Testing->SelfTests
options.  This is a requirement to add the module to the
2014 NAM-IC Module Tutorial Contest. The default self-test
will be adjusted to test the features of this module.