ENH: Adds rasToXY, updates rasPoint and testScores
The paintTool.paintAddPoint(x, y) requires and x,y
input that can change with width or height changes
of the Slicer application. The RAS point does not
change, so I added a function to convert from RAS
to XY coordinates in order to paint in the same
location each time the test is ran, regardless of
the dimensions of the Slicer Application. I also
updated the testScores dictionary to represent the
new scores calculated at the new RAS points.